# docs/swagger.yaml
basePath: /
definitions:
...
  response.ErrorResponse:
    description: Error response
    properties:
      message:
        example: error message
        type: string
      success:
        example: false
        type: boolean
    type: object
  response.SuccessResponse:
    description: Success response
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:5000
info:
  contact:
    email: support@go-sosmed.com
    name: API Support
  description: API for Golang Social Media Application
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GO-SOSMED API
  version: "1.0"
paths:
...
  /api/follow/{following_id}:
    delete:
      consumes:
      - application/json
      description: Unfollow a user you're currently following
      parameters:
      - description: User ID to unfollow
        in: path
        name: following_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unfollow a user
      tags:
      - Follow
    post:
      consumes:
      - application/json
      description: Follow another user
      parameters:
      - description: User ID to follow
        in: path
        name: following_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Follow a user
      tags:
      - Follow
  /api/follow/me/followers:
    get:
      consumes:
      - application/json
      description: Get list of users following the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get followers
      tags:
      - Follow
  /api/follow/me/following:
    get:
      consumes:
      - application/json
      description: Get list of users the authenticated user is following
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get following
      tags:
      - Follow
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"